quadtree vertex array indices now uses GL_UNSIGNED_SHORT instead of GL_UNSIGNED_INT

+++ quadtree.h
@@ -78,7 +78,7 @@ struct quadsquare {
     static GLuint TexId[MAX_TERR_TYPES];
     static GLuint EnvmapTexId;
 
-    static GLuint *VertexArrayIndices;
+    static GLushort *VertexArrayIndices;
     static GLuint VertexArrayCounter;
     static GLuint VertexArrayMinIdx;
     static GLuint VertexArrayMaxIdx;
